DATABASETOOL: fixed msvc issue

+++ src/modules/core/String.h
@@ -75,7 +75,7 @@ inline bool toBool(const std::string& str) {
 }
 
 inline float toFloat(const std::string& str) {
-	return SDL_atof(str.c_str());
+	return (float)SDL_atof(str.c_str());
 }
 
 extern void splitString(const std::string& string, std::vector<std::string>& tokens, const std::string& delimiters = " \t\r\n\f\v");
